# ci: Improve gadget building time 