- Move the pulling logic to a different step on the makefile to avoid pulling the gadget builder for each gadget.
- Fix a bug found while trying the above 

### Testing 

0. Be sure you have the latest gadget-builder image available:

```bash 
$ docker pull ghcr.io/inspektor-gadget/gadget-builder:main
main: Pulling from inspektor-gadget/gadget-builder
Digest: sha256:dac71c0de0d83ecc8e7785688548fc795a1e8c374e4dfa2827fe982dd99f8584
Status: Image is up to date for ghcr.io/inspektor-gadget/gadget-builder:main
ghcr.io/inspektor-gadget/gadget-builder:main
```

### Before this PR 

Building with a single core on my machine takes around 2 minutes 

```bash
$ time make -C gadgets
...

real    1m52,879s
user    0m0,411s
sys     0m0,458s
```

### After this PR 

Building with a single core only takes around 1m20s.

```bash 
$ time make -C gadgets

real    1m18,537s
user    0m0,648s
sys     0m0,521s
```

EDIT: I initially opened this hoping the time used in the CI will be lower, but it seems to be the same, probably because the latency between github runners and ghcr is very small, hence trying to pull the image multiple times doesn't have a big impact as when running it locally. Please check it on your systems and let me know if you see any improvement.