@mauriciovasquezbernal 2025-05-23
As it doesn't harm anybody and all tests are green, I'll move ahead and merge it.